Move conditional imports to _compat module
